requirements.txt
numpy
python_dotenv
qi